requirements.txt
crewai>=0.20.0
openai>=1.5.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
jsonschema>=4.19.0
colorama>=0.4.6 
onnxruntime
langchain>=0.1.0
langchain-openai>=0.0.3
